Implement tasks ordering following DDD principles
In domain model, Tasks are stored just as an [ordered] list inside
a Section object, which makes it easy to operate with them. When
(and only when) persisted with SQLAlchemyRepositories, the
`task_model.index` appears as a way to keep the order.

pytest => ok (but tests should be re-read and extended)
ruff => ok
mypy => not ok

Also, I added a funny button to the frontend that shuffles tasks in
a section (I guess sometimes it can even be useful 🎩).